+++ swig/cpp/src/Session.cpp
@@ -27,7 +27,7 @@
 #include "Struct.h"
 #include "Internal.h"
 #include "Tree.h"
-#include "Sysrepo.h"
+#include "Sysrepo.hpp"
 #include "Connection.h"
 #include "Session.h"
