+++ swig/lua/libsysrepoLua.i
@@ -27,7 +27,7 @@
 #include <vector>
 #include <memory>
 
-#include "Sysrepo.h"
+#include "Sysrepo.hpp"
 #include "Struct.h"
 #include "Session.h"
