+++ swig/python2/libsysrepoPython2.i
@@ -25,7 +25,7 @@
 #include <vector>
 #include <memory>
 
-#include "Sysrepo.h"
+#include "Sysrepo.hpp"
 #include "Struct.h"
 #include "Session.h"
